BE work on purchase; FE mapping stocks

+++ scripts/scripts.sql
@@ -7,8 +7,8 @@ CREATE TABLE `users` (
   `password` varchar(500) DEFAULT NULL,
   `display_name` varchar(45) DEFAULT NULL,
   `email` varchar(45) DEFAULT NULL,
-  `cash` int(45) DEFAULT NULL,
-  `invested_balance` int(45) DEFAULT NULL,
+  `cash` float(45) DEFAULT NULL,
+  `invested_balance` float(45) DEFAULT NULL,
   `enabled` tinyint(1) DEFAULT NULL,
 
   PRIMARY KEY (`username`)
@@ -33,7 +33,7 @@ CREATE TABLE `stock` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `username` varchar(55),
   `symbol` varchar(20),
-  `purchase_price` int(45) DEFAULT 0,
+  `purchase_price` float(45) DEFAULT 0,
   `purchase_date` timestamp (6) DEFAULT NULL,
   `quantity` int(100) DEFAULT NULL,
   `user` varchar(55),
